voice-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
strands-agents[bidi-all]==1.29.0
boto3>=1.35.0
httpx>=0.27.0
python-dotenv>=1.0.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-instrumentation-httpx>=0.46b0
